Work in progress: Simple product support.
