+++ composer.json
@@ -3,7 +3,8 @@
     "description": "Change product image on hover on category pages",
     "require": {
         "php": "^7.1||^8.1",
-        "magento/module-catalog": "*"
+        "magento/module-catalog": "*",
+        "magento/module-swatches": "*"
     },
     "type": "magento2-module",
     "license": [
