+++ view/frontend/layout/catalog_category_view.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <body>
+        <referenceContainer name="category.product.list.additional">
+            <block class="Magento\Framework\View\Element\Template"
+                   name="dmatthew.catalog.additional.hoverimage"
+                   template="Dmatthew_HoverImage::hover_image.phtml"/>
+        </referenceContainer>
+    </body>
+</page>
