--- view/frontend/requirejs-config.js
@@ -1,9 +1,0 @@
-var config = {
-    config: {
-        mixins: {
-            'Magento_Swatches/js/swatch-renderer': {
-                'Dmatthew_HoverImage/js/swatch-renderer-mixin': true
-            }
-        }
-    }
-};
