+++ view/frontend/templates/hover_image.phtml
@@ -1,0 +1,10 @@
+<?php
+/** @var Magento\Framework\View\Element\Template $block */
+?>
+<script type="text/x-magento-init">
+    {
+        "[data-hover-image]": {
+            "Dmatthew_HoverImage/js/hover_image": {}
+        }
+    }
+</script>
